bsod: fix sign compare warning
This commit fixes the following warning:
bsod.cpp:313:20: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
  for (cnt = 1; cnt < size; ++cnt)

+++ main/bsod.cpp
@@ -306,7 +306,7 @@ void print_backtrace()
 {
 	void *array[15];
 	size_t size;
-	int cnt;
+	size_t cnt;
 
 	size = backtrace(array, 15);
 	eLog(lvlFatal, "Backtrace:");
